Fix All Supabase, OpenAI, and Receipt Processing Integrations
