+++ src/lib/supabase.ts
@@ -11,8 +11,7 @@ export const supabase = createClient(supabaseUrl, supabaseAnonKey, {
   auth: {
     autoRefreshToken: true,
     persistSession: true,
-    detectSessionInUrl: true,
-    flowType: 'pkce'
+    detectSessionInUrl: true
   }
 })
 
@@ -772,5 +771,123 @@ export const testOpenAIConnection = async () => {
   } catch (err) {
     console.error('OpenAI API connection test error:', err);
     return false;
+  }
+}
+
+// GPT-powered receipt data extraction
+export const extractReceiptDataWithGPT = async (extractedText: string) => {
+  try {
+    const openaiApiKey = import.meta.env.VITE_OPENAI_API_KEY;
+    
+    if (!openaiApiKey) {
+      throw new Error('OpenAI API key not configured');
+    }
+
+    console.log('Extracting receipt data with GPT...');
+
+    const prompt = `You are a reliable AI that extracts structured receipt details from plain text.
+Given raw text from a receipt, extract the following fields and return ONLY valid JSON.
+If a field is missing, set its value to null. No extra explanation or text outside the JSON.
+
+Required JSON fields:
+- product_description (string): Main product or service purchased
+- brand_name (string): Brand or manufacturer name
+- store_name (string): Name of the store/merchant
+- purchase_location (string): Store location/address
+- purchase_date (string): Date in YYYY-MM-DD format
+- amount (number): Total amount paid (numeric only, no currency symbol)
+- warranty_period (string): Warranty duration (e.g., "1 year", "6 months")
+- extended_warranty (string): Extended warranty info if any
+- model_number (string): Product model number if available
+- country (string): Country where purchase was made
+
+Receipt text:
+${extractedText}
+
+Return only valid JSON:`;
+
+    const response = await fetch('https://api.openai.com/v1/chat/completions', {
+      method: 'POST',
+      headers: {
+        'Authorization': `Bearer ${openaiApiKey}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        model: 'gpt-3.5-turbo',
+        messages: [
+          {
+            role: 'system',
+            content: 'You are a precise data extraction assistant. Return only valid JSON with the requested fields.'
+          },
+          {
+            role: 'user',
+            content: prompt
+          }
+        ],
+        max_tokens: 500,
+        temperature: 0.1,
+      }),
+    });
+
+    if (!response.ok) {
+      const errorData = await response.text();
+      console.error('OpenAI API error:', response.status, errorData);
+      throw new Error(`OpenAI API error: ${response.status}`);
+    }
+
+    const data = await response.json();
+    const content = data.choices[0]?.message?.content;
+
+    if (!content) {
+      throw new Error('No response from OpenAI');
+    }
+
+    // Clean the response to extract JSON
+    const jsonMatch = content.match(/\{[\s\S]*\}/);
+    if (!jsonMatch) {
+      throw new Error('No valid JSON found in response');
+    }
+
+    const extractedData = JSON.parse(jsonMatch[0]);
+    
+    // Validate and clean the extracted data
+    const cleanedData = {
+      product_description: extractedData.product_description || 'Receipt Item',
+      brand_name: extractedData.brand_name || 'Unknown Brand',
+      store_name: extractedData.store_name || null,
+      purchase_location: extractedData.purchase_location || null,
+      purchase_date: extractedData.purchase_date || new Date().toISOString().split('T')[0],
+      amount: extractedData.amount && typeof extractedData.amount === 'number' ? extractedData.amount : null,
+      warranty_period: extractedData.warranty_period || '1 year',
+      extended_warranty: extractedData.extended_warranty || null,
+      model_number: extractedData.model_number || null,
+      country: extractedData.country || 'United States'
+    };
+
+    // Ensure amount is a number or null
+    if (cleanedData.amount && typeof cleanedData.amount === 'string') {
+      const numericAmount = parseFloat(cleanedData.amount.replace(/[^0-9.]/g, ''));
+      cleanedData.amount = isNaN(numericAmount) ? null : numericAmount;
+    }
+
+    // Ensure date is in correct format
+    if (cleanedData.purchase_date && !cleanedData.purchase_date.match(/^\d{4}-\d{2}-\d{2}$/)) {
+      const date = new Date(cleanedData.purchase_date);
+      if (!isNaN(date.getTime())) {
+        cleanedData.purchase_date = date.toISOString().split('T')[0];
+      } else {
+        cleanedData.purchase_date = new Date().toISOString().split('T')[0];
+      }
+    }
+
+    console.log('GPT extraction successful:', cleanedData);
+    return { data: cleanedData, error: null };
+
+  } catch (error: any) {
+    console.error('GPT extraction error:', error);
+    return { 
+      data: null, 
+      error: { message: error.message || 'Failed to extract data with GPT' } 
+    };
   }
 }
